Move EPANET Toolkit to new location, start removing unused linkage functions.
