+++ wntr/epanet/__init__.py
@@ -3,5 +3,5 @@ The wntr.epanet package provides EPANET2 compatibility functions for WNTR.
 """
 from .io import InpFile  #, BinFile, HydFile, RptFile
 from .util import FlowUnits, MassUnits, HydParam, QualParam, EN
-import wntr.epanet.pyepanet
+import wntr.epanet.toolkit
