--- wntr/epanet/pyepanet/Externals
@@ -1,1 +1,0 @@
-^/BASE/tags/2.00.12/Bin data
